[CLI] remove cli_commands module.
