+++ cli/src/pcluster/cli/commands/commands.py
@@ -1,0 +1,29 @@
+# Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance
+# with the License. A copy of the License is located at
+#
+# http://aws.amazon.com/apache2.0/
+#
+# or in the "LICENSE.txt" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
+# OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
+# limitations under the License.
+
+# pylint: disable=R0801
+
+# flake8: noqa
+
+from pcluster.cli.commands.cluster_logs import (
+    GetClusterLogEventsCommand,
+    ListClusterLogsCommand,
+    ExportClusterLogsCommand,
+)
+from pcluster.cli.commands.image_logs import (
+    GetImageLogEventsCommand,
+    ListImageLogsCommand,
+    ExportImageLogsCommand,
+)
+from pcluster.cli.commands.configure.command import ConfigureCommand
+from pcluster.cli.commands.dcv_connect import DcvConnectCommand
+from pcluster.cli.commands.ssh import SshCommand
+from pcluster.cli.commands.version import VersionCommand
