+++ cli/src/pcluster/cli/commands/image_logs.py
