+++ cli/src/pcluster/cli/commands/version.py
