+++ cli/src/pcluster/models/compute_fleet_status_manager.py
