+++ cli/tests/pcluster/cli/configure/test_pcluster_configure/test_with_region_arg/output.txt
