+++ cli/tests/pcluster/cli/test_commands.py
