+++ cli/tests/pcluster/cli/test_entrypoint/TestParallelClusterCli/test_helper/pcluster-help.txt
@@ -1,5 +1,5 @@
 usage: pcluster [-h]
-                {list-clusters,create-cluster,delete-cluster,describe-cluster,update-cluster,describe-compute-fleet,update-compute-fleet,delete-cluster-instances,describe-cluster-instances,list-images,build-image,delete-image,describe-image,describe-official-images,configure,dcv-connect,export-cluster-logs,get-cluster-log-events,list-cluster-logs,ssh,export-image-logs,get-image-log-events,list-image-logs,version}
+                {list-clusters,create-cluster,delete-cluster,describe-cluster,update-cluster,describe-compute-fleet,update-compute-fleet,delete-cluster-instances,describe-cluster-instances,list-images,build-image,delete-image,describe-image,describe-official-images,configure,dcv-connect,export-cluster-logs,export-image-logs,get-cluster-log-events,get-image-log-events,list-cluster-logs,list-image-logs,ssh,version}
                 ...
 
 pcluster is the AWS ParallelCluster CLI and permits launching and management
@@ -9,7 +9,7 @@ optional arguments:
   -h, --help            show this help message and exit
 
 COMMANDS:
-  {list-clusters,create-cluster,delete-cluster,describe-cluster,update-cluster,describe-compute-fleet,update-compute-fleet,delete-cluster-instances,describe-cluster-instances,list-images,build-image,delete-image,describe-image,describe-official-images,configure,dcv-connect,export-cluster-logs,get-cluster-log-events,list-cluster-logs,ssh,export-image-logs,get-image-log-events,list-image-logs,version}
+  {list-clusters,create-cluster,delete-cluster,describe-cluster,update-cluster,describe-compute-fleet,update-compute-fleet,delete-cluster-instances,describe-cluster-instances,list-images,build-image,delete-image,describe-image,describe-official-images,configure,dcv-connect,export-cluster-logs,export-image-logs,get-cluster-log-events,get-image-log-events,list-cluster-logs,list-image-logs,ssh,version}
     list-clusters       Retrieve the list of existing clusters.
     create-cluster      Create a managed cluster in a given region.
     delete-cluster      Initiate the deletion of a cluster.
@@ -38,15 +38,15 @@ COMMANDS:
     export-cluster-logs
                         Export the logs of the cluster to a local tar.gz
                         archive by passing through an Amazon S3 Bucket.
-    get-cluster-log-events
-                        Retrieve the events of a log stream of the cluster.
-    list-cluster-logs   List the log streams associated to a cluster.
-    ssh                 Connects to the head node instance using SSH.
     export-image-logs   Export the logs of the image builder stack to a local
                         tar.gz archive by passing through an Amazon S3 Bucket.
+    get-cluster-log-events
+                        Retrieve the events of a log stream of the cluster.
     get-image-log-events
                         Retrieve the events of a log stream of the image.
+    list-cluster-logs   List the log streams associated to a cluster.
     list-image-logs     List the log streams associated to an image id.
+    ssh                 Connects to the head node instance using SSH.
     version             Displays the version of AWS ParallelCluster.
 
 For command specific flags, please run: "pcluster [command] --help"
