+++ cli/tests/pcluster/cli/test_entrypoint/TestParallelClusterCli/test_no_command/pcluster-command-error.txt
@@ -1,4 +1,4 @@
 usage: pcluster [-h]
-                {list-clusters,create-cluster,delete-cluster,describe-cluster,update-cluster,describe-compute-fleet,update-compute-fleet,delete-cluster-instances,describe-cluster-instances,list-images,build-image,delete-image,describe-image,describe-official-images,configure,dcv-connect,export-cluster-logs,get-cluster-log-events,list-cluster-logs,ssh,export-image-logs,get-image-log-events,list-image-logs,version}
+                {list-clusters,create-cluster,delete-cluster,describe-cluster,update-cluster,describe-compute-fleet,update-compute-fleet,delete-cluster-instances,describe-cluster-instances,list-images,build-image,delete-image,describe-image,describe-official-images,configure,dcv-connect,export-cluster-logs,export-image-logs,get-cluster-log-events,get-image-log-events,list-cluster-logs,list-image-logs,ssh,version}
                 ...
 pcluster: error: the following arguments are required: operation
