+++ cli/tests/pcluster/cli/test_export_cluster_logs.py
@@ -74,7 +74,7 @@ class TestExportClusterLogsCommand:
         ],
     )
     def test_execute(self, mocker, capsys, set_env, assert_out_err, run_cli, args):
-        export_logs_mock = mocker.patch("pcluster.cli.commands.cluster.Cluster.export_logs")
+        export_logs_mock = mocker.patch("pcluster.cli.commands.cluster_logs.Cluster.export_logs")
         set_env("AWS_DEFAULT_REGION", "us-east-1")
 
         command = BASE_COMMAND + self._build_cli_args({**REQUIRED_ARGS, **args})
