+++ cli/tests/pcluster/cli/test_get_cluster_log_events.py
@@ -116,10 +116,10 @@ class TestGetClusterLogEventsCommand:
             )
         ] * 2 + [LogStream(FAKE_NAME, "logstream", {})]
         get_cluster_log_events_mock = mocker.patch(
-            "pcluster.cli.commands.cluster.Cluster.get_log_events", side_effect=mocked_result
+            "pcluster.cli.commands.cluster_logs.Cluster.get_log_events", side_effect=mocked_result
         )
         set_env("AWS_DEFAULT_REGION", "us-east-1")
-        mocker.patch("pcluster.cli.commands.cluster.time.sleep")  # so we don't actually have to wait
+        mocker.patch("pcluster.cli.commands.cluster_logs.time.sleep")  # so we don't actually have to wait
 
         command = BASE_COMMAND + self._build_cli_args({**REQUIRED_ARGS, **args})
